pyproject.toml
[project]
name = "agentic-workflows-tools"
version = "0.1.0"
description = "Helper tooling for the agentic-workflows repository"
requires-python = ">=3.14,<3.15"
dependencies = [
  "PyYAML>=6.0.2,<7",
]

[tool.uv]
package = false
